requirements.txt
streamlit
pandas
requests
python-dotenv
crewai>=0.30.0
crewai-tools
langchain-google-genai
pydantic>=2.4.1